Fixes #123 - Avoid caching song_id form parameter
Restructure the add to playlist form so that instead of a list of buttons, each generating a form, we wrap the playlist buttons in a single form and each playlist button changes the form action when clicked to the proper playlist url.

This allows caching each button and fixes the bug where the song_id was cached with the html for each button.